social_network/requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
cloudinary==1.25.0
Django==3.1.7
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
idna==2.10
Markdown==3.3.4
Pillow==8.2.0
psycopg2==2.8.6
PyJWT==2.0.1
python-dotenv==0.17.0
pytz==2021.1
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.4